Refactor: Improve Dockerfile, README, and JaCoCo exclusions
- Refactored Dockerfile to use multi-stage build for smaller images.
- Updated README.md to reflect Dockerfile changes and user registration endpoint.
- Centralized JaCoCo exclusions for DTO and Entity classes into jacoco_excludes.gradle.
- Updated build.gradle to use jacoco_excludes.gradle for JaCoCo configuration.
