+++ src/main/java/com/example/iamsystem/permission/PermissionService.java
@@ -2,6 +2,10 @@ package com.example.iamsystem.permission;
 
 import com.example.iamsystem.exception.DataNotFoundException;
 import com.example.iamsystem.exception.PermissionAlreadyExistsException;
+import com.example.iamsystem.permission.model.Permission;
+import com.example.iamsystem.permission.model.PermissionAction;
+import com.example.iamsystem.permission.model.PermissionDto;
+import com.example.iamsystem.permission.model.PermissionMapper;
 import com.example.iamsystem.user.model.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.factory.Mappers;
