+++ src/main/java/com/example/iamsystem/role/RoleRepository.java
@@ -1,5 +1,6 @@
 package com.example.iamsystem.role;
 
+import com.example.iamsystem.role.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
