+++ src/main/java/com/example/iamsystem/role/model/Role.java
@@ -1,39 +1,51 @@
-package com.example.iamsystem.permission;
+package com.example.iamsystem.role.model;
 
-import jakarta.persistence.Column;
+import com.example.iamsystem.permission.model.Permission;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import java.io.Serializable;
 import java.time.Instant;
+import java.util.HashSet;
 import java.util.Objects;
+import java.util.Set;
 
 @Entity
-@Table(name = "permissions")
+@Table(name = "roles")
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Permission implements Serializable {
+public class Role implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(name = "service_name")
-    private String serviceName;
-    @Enumerated(value = EnumType.STRING)
-    private PermissionAction action;
+    private String name;
     private String description;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(
+            name = "role_permissions",
+            joinColumns = @JoinColumn(name = "role_id", referencedColumnName = "id"),
+            inverseJoinColumns = @JoinColumn(name = "permission_id", referencedColumnName = "id")
+    )
+    @OnDelete(action = OnDeleteAction.CASCADE)
+    private Set<Permission> permissions = new HashSet<>();
     @CreationTimestamp
     private Instant createdAt;
     @UpdateTimestamp
@@ -43,8 +55,8 @@ public class Permission implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Permission that = (Permission) o;
-        return Objects.equals(getId(), that.getId());
+        Role role = (Role) o;
+        return Objects.equals(getId(), role.getId());
     }
 
     @Override
