+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=User Management App
+spring.application.name=IAM System Application
 spring.datasource.url=jdbc:mysql://localhost:3366/iam_system_db
 spring.datasource.username=root
 spring.datasource.password=password
